♻️ refactor:  GitHubAppのファイル名を変更
Issue #145
